move targets into namespace packages
